requirements.txt
# ── Interface ────────────────────────────────────────────────
streamlit>=1.32.0

# ── Traitement d'images ──────────────────────────────────────
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0

# ── Données ──────────────────────────────────────────────────
pandas>=2.0.0

# ── Machine Learning ─────────────────────────────────────────
scikit-learn>=1.3.0
scipy>=1.11.0

# ── Deep Learning (CPU suffit, GPU optionnel) ─────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── SAM — Segment Anything Model (Meta) ──────────────────────
# Installer manuellement si pip install échoue :
#   pip install git+https://github.com/facebookresearch/segment-anything.git
segment-anything>=1.0
